feat: add authorize module in middlewares
