+++ faster_sam/middlewares/remove_path.py
@@ -1,8 +1,3 @@
-import json
-from http import HTTPStatus
-from typing import Dict, Optional
-
-from botocore.client import BaseClient
 from starlette.middleware.base import BaseHTTPMiddleware, RequestResponseEndpoint
 from starlette.requests import Request
 from starlette.responses import Response
@@ -58,75 +53,3 @@ class RemovePathMiddleware(BaseHTTPMiddleware):
 
         request.scope["path"] = request.scope["path"].replace(self.path, "", 1)
         return await call_next(request)
-
-
-class LambdaAuthorizerMiddleware(BaseHTTPMiddleware):
-    """
-    Invoke Lambda function in AWS to authorize apis
-
-    e.g
-
-    This example apply the middleware to authorize an app.
-
-    >>> app = FastAPI()
-    >>> app.add_middleware(LambdaAuthorizer, arn="arn:aws:lambda:region:id:function:name")
-
-    Parameters
-    ----------
-        app : ASGIApp
-            Application instance the middleware is being registered to.
-        arn : str
-            The amazon resource name for the Lambda will be invoked.
-        client: BaseClient
-            Client Lambda object
-    """
-
-    def __init__(self, app: ASGIApp, arn: str, client: Optional[BaseClient] = None) -> None:
-        """
-        Initializes the LambdaAuthorizer.
-        """
-        super().__init__(app, self.dispatch)
-        self.arn = arn
-        self.client = client
-
-    async def dispatch(self, request: Request, call_next: RequestResponseEndpoint) -> Response:
-        """
-        Authorize or deny a request based on Authorization field.
-
-        Parameters
-        ----------
-        request : Request
-            The incoming request.
-        call_next : RequestResponseEndpoint
-            Next middleware or endpoint on the execution stack
-
-        Returns
-        -------
-        Response
-            The response generated by the middleware.
-        """
-
-        response = self.invoke_lambda(request)
-
-        if not response:
-            return Response(
-                content=json.dumps({"message": "Something went wrong. Try again"}),
-                status_code=HTTPStatus.INTERNAL_SERVER_ERROR.value,
-            )
-        return await call_next(request)
-
-    def invoke_lambda(self, request: Request) -> Dict[str, any]:
-        """
-        Invoke a Lambda using arn from AWS.
-
-        Parameters
-        ----------
-        request : Request
-            The incoming request.
-
-        Returns
-        -------
-        Response
-            The response generated by AWS Lambda invoked.
-        """
-        return {}
