+++ pyproject.toml
@@ -20,3 +20,6 @@ dependencies = [
     "fastapi",
     "PyYAML == 6.0.1",
 ]
+
+[project.optional-dependencies]
+AUTH = ["boto3 == 1.34.19"]
